+++ src/main/java/seedu/superta/model/assignment/GradeBook.java
@@ -12,14 +12,14 @@ import seedu.superta.model.student.StudentId;
  * Model for a grade book.
  */
 public class GradeBook {
-    private final HashMap<StudentId, Integer> internalHashmap = new HashMap<>();
+    private final HashMap<StudentId, Double> internalHashmap = new HashMap<>();
 
     /**
      * Adds a grade to the internal hashmap
      * @param stId the student ID.
      * @param grade the grade value
      */
-    public void addGrade(StudentId stId, int grade) {
+    public void addGrade(StudentId stId, Double grade) {
         internalHashmap.put(stId, grade);
     }
 
@@ -28,7 +28,7 @@ public class GradeBook {
      * @param stId the student ID.
      * @return his grade for this assignment
      */
-    public int getGradeFor(StudentId stId) {
+    public Double getGradeFor(StudentId stId) {
         return internalHashmap.get(stId);
     }
 
@@ -54,7 +54,7 @@ public class GradeBook {
      * Method to streamify this object. Also, orders it in lexicographical order of Student IDs.
      * @return a Stream of entries in lexicographical order.
      */
-    public Stream<Map.Entry<StudentId, Integer>> stream() {
+    public Stream<Map.Entry<StudentId, Double>> stream() {
         return internalHashmap.entrySet().stream()
             .sorted(Comparator.comparing(o -> o.getKey().toString()));
     }
